+++ R/boostrap_clust_validation.R
@@ -72,9 +72,9 @@ boot_group_validation <- function(object, nsim = 1000, maxiter = 1000, tol = 0.0
     }else{
       dataset <- lapply(1:ncol(object$Data), function(i){
         rast <- object$rasters[[1]]
-        vec <- rep(NA, times = raster::ncell(rast))
+        vec <- rep(NA, times = terra::ncell(rast))
         vec[object$missing] <- object$Data[,i]
-        raster::values(rast) <- vec
+        terra::values(rast, mat = FALSE) <- vec
         return(rast)
       })
       wdata <- calcWdataRaster(object$window, dataset, object$lag_method, object$missing)
@@ -242,9 +242,9 @@ boot_group_validation.mc <- function(object, nsim = 1000, maxiter = 1000, tol = 
     }else{
       dataset <- lapply(1:ncol(object$Data), function(i){
         rast <- object$rasters[[1]]
-        vec <- rep(NA, times = raster::ncell(rast))
+        vec <- rep(NA, times = terra::ncell(rast))
         vec[object$missing] <- object$Data[,i]
-        raster::values(rast) <- vec
+        terra::values(rast, mat = FALSE) <- vec
         return(rast)
       })
       wdata <- calcWdataRaster(object$window, dataset, object$lag_method, object$missing)